route and bus details fetch
