+++ static/css/details.css
@@ -1,0 +1,152 @@
+
+    /* Reset styles */
+* {
+    margin: 0;
+    padding: 0;
+    font-family: sans-serif;
+}
+
+/* Body styles */
+body {
+    overflow: auto;
+}
+
+/* Main container styles */
+.main {
+    background-image: linear-gradient(rgba(0, 0, 0, 0.40), rgba(0, 0, 0, 0.40)), url('/media/p.png');
+    height: 240vh;
+    width: 100%;
+    background-position: center;
+    background-size: cover;
+    position: relative;
+}
+
+/* Logo container styles */
+.logo-container {
+    position: absolute;
+    top: 40px;
+    left: 50px;
+    z-index: 1000;
+}
+
+/* Logo styles */
+.logo {
+    max-width: 219px;
+    height: auto;
+}
+
+/* Navigation bar styles */
+.navbar {
+    width: 85%;
+    margin: auto;
+    padding: 38px 0;
+    display: flex;
+    align-items: center;
+    justify-content: flex-end;
+    position: absolute;
+    top: 0;
+    left: 0;
+    right: 0;
+    padding-left: 140px;
+}
+
+/* Navigation link styles */
+.navbar ul li {
+    list-style: none;
+    display: inline-block;
+    margin: 10px 20px;
+    position: relative;
+    font-size: 20.3px;
+    text-shadow: 0 4px 8px rgb(29, 29, 29);
+}
+
+/* Navigation link hover styles */
+.navbar ul li a:hover {
+    color: #04AA6D;
+}
+/* Styling for navigation links */
+.navbar ul li a {
+    text-decoration: none;
+    color: #f2f2f2;
+    transition: color 0.5s;
+}
+
+/* Navigation link underline styles */
+.navbar ul li::after {
+    content: '';
+    height: 3px;
+    width: 0;
+    background: #14b97c;
+    position: absolute;
+    left: 0;
+    bottom: -10px;
+    transition: 0.5s;
+}
+
+/* Navigation link underline on hover styles */
+.navbar ul li:hover::after {
+    width: 100%;
+}
+
+.black-panel2 {
+    position: absolute;
+    top: 300px;
+    left: 108px;
+    width: 50%;
+    height: 40%;
+    background-color: rgba(0, 0, 0, 0.5);
+    border-radius: 12px;
+    border: 1.3px solid white;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+}
+/* Heading styles */
+.panel-heading {
+    color: white;
+    font-size: 23px;
+    text-align: center;
+    margin-top: -650px;
+    position: absolute;
+}
+.entry-box {
+            margin: 20px;
+            padding: 15px;
+            width: 70%;
+            border: 1px solid #ccc;
+            border-radius: 4px;
+            box-sizing: border-box;
+            color: #ffff;
+            font: 100px;
+            background-color:  rgba(255, 253, 253, 0.221);;
+        }
+    .next-button {
+    background-color: #04AA6D;
+    color: #ffffff;
+    border: none;
+    padding: 10px 10px;
+    font-size: 18px;
+    cursor: pointer;
+    width: 150px;
+    height: 40px;
+    border: #ffffff;
+    border-radius: 5px;
+    position: absolute; /* Set the position to absolute */
+    bottom: 18px;
+    right: 320px; /* Adjusted right positioning */
+    font: bold;
+}
+
+.next-button:hover {
+    background-color: #00e18f;
+}
+.input-label {
+            color: white;
+            font-size: 17px;
+            margin-bottom: -10px;
+            margin-right: 320px;
+        }
+
+
+
